refactor: fix react-refresh warnings for `src/contexts`
