+++ src/contexts/GustoApiProvider/index.ts
@@ -1,0 +1,2 @@
+/** @deprecated Import from `GustoProvider` instead */
+export * from '../GustoProvider'
