--- src/contexts/GustoApiProvider/index.tsx
@@ -1,3 +1,0 @@
-export * from './GustoProvider'
-export { SDKI18next } from './SDKI18next'
-export * from './GustoProviderCustomUIAdapter'
