+++ src/contexts/LocaleProvider/index.ts
@@ -1,0 +1,2 @@
+export { LocaleProvider } from './LocaleProvider'
+export { useLocale, type LocaleProps } from './useLocale'
