--- src/contexts/LocaleProvider/index.tsx
@@ -1,1 +1,0 @@
-export * from './LocaleProvider'
